+++ libmov/source/mov-trun.c
@@ -84,12 +84,12 @@ int mov_read_trun(struct mov_t* mov, const struct mov_box_t* box)
 	return mov_buffer_error(&mov->io);
 }
 
-size_t mov_write_trun(const struct mov_t* mov, size_t from, size_t count, uint32_t moof)
+size_t mov_write_trun(const struct mov_t* mov, uint32_t from, uint32_t count, uint32_t moof)
 {
     uint32_t flags;
 	uint32_t delta;
 	uint64_t offset;
-	size_t size, i;
+	uint32_t size, i;
 	const struct mov_sample_t* sample;
 	const struct mov_track_t* track = mov->track;
